mkv: factor out some modules
